* new: make red/green/blue for gray * fix: 332 container is not uint8_t * fix: gray -> cmy * fix: gray -> rgb * fix: rgb make for unsigned containers

+++ src/color/_internal/utility/container/cner332.hpp
@@ -22,8 +22,8 @@ namespace color
 
               typedef index_name     index_type;
 
-              typedef std::uint8_t  value_type;
-              typedef std::uint16_t  instance_type;
+              typedef std::uint8_t   value_type;
+              typedef std::uint8_t   instance_type;
 
               typedef instance_type const      const_type;
               typedef instance_type const&     return_const_type;
